Updated Makefiles
This diff updates the examples makefile to make all of the tb_* files in the
directory, as well as includes a very rudimentary (to be completed later)
correctness checking. Currently, it only checks that the files have been
successfully compiled. By default, make runs Phase-II compilation. To run
Phase-I compilation, run make DEBUG=1.