Updated heat_1D_NP to work under new interface
There are a couple of interesting quirks with the new interface:
1. Even though the kernel we wish to use is not tiled, one must still call
Register_Tile_Kernels, as if one doesn't you get the "must register Shape
before Array", even if you are indeed calling Register_Shape before
Register_Array.
2. Calling run with a single argument (namely, the time steps) triggers
Run_Obase_Unroll_T, but calling run with a second argument (being the kernel)
doesn't. Not really sure if this is intentional or not, but it seems somewhat
odd. Particularly, given that I must register the kernel, passing in the kernel
again as an argument to avoid triggering Run_Obase_Unroll_T seems a bit odd.