Partially updated tb_heat_2D_NP_zero
It now compiles and runs, though the reference and pochoir implementations
don't produce the same results. This benchmark seems to work fine in the master
branch, so I'm not entirely sure as to what's off. When removing the heat_2D_fn
argument from heat_2D.Run (triggering the plan generator) the generator
produces uncompilable C++ code.